Added some print statements for debugging. We can see that there are some serious offsets now since the offsets are being calculated before the inputs xfos are set due to the fact that they don't resolve until being built in the DCC's.
TODO: Figure out how to handle this. Previously we had maintainOffset turned on for inputs. Now this is turned off to ensure inputs match the outputs they are connected to. We may need to traverse the graph and resolve the inputs / outputs in order but this is tricky as you can have cyclic connections (leg > foot > leg).